+++ src/GitHub_Updater/Additions.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * GitHub Updater
+ *
+ * @package   GitHub_Updater
+ * @author    Andy Fragen
+ * @license   GPL-2.0+
+ * @link      https://github.com/afragen/github-updater
+ */
 
 namespace Fragen\GitHub_Updater;