+++ src/GitHub_Updater/Base.php
@@ -252,7 +252,8 @@ class Base {
 			} );
 
 			// Run GitHub Updater upgrade functions.
-			new GHU_Upgrade();
+			$upgrade = new GHU_Upgrade();
+			$upgrade->run();
 
 			// Ensure transient updated on plugins.php and themes.php pages.
 			add_action( 'admin_init', array( &$this, 'admin_pages_update_transient' ) );