+++ src/GitHub_Updater/Basic_Auth_Loader.php
@@ -136,8 +136,8 @@ class Basic_Auth_Loader {
 			'private'       => false,
 		);
 
-		$slug  = isset( $_REQUEST['rollback'], $_REQUEST['plugin'] ) ? dirname( $_REQUEST['plugin'] ) : false;
-		$slug  = isset( $_REQUEST['rollback'], $_REQUEST['theme'] ) ? $_REQUEST['theme'] : $slug;
+		$slug  = isset( $_REQUEST['plugin'] ) ? dirname( $_REQUEST['plugin'] ) : false;
+		$slug  = isset( $_REQUEST['theme'] ) ? $_REQUEST['theme'] : $slug;
 		$slug  = isset( $_REQUEST['slug'] ) ? $_REQUEST['slug'] : $slug;
 		$repos = isset( $_REQUEST )
 			? array_merge(